refactor(front): split flow edit sheet into modular edit-panel components
