--- apps/ferrisquote-webapp/src/pages/flows/ui/flow-edit-sheet.tsx
@@ -1,1316 +1,0 @@
-import { useState, useEffect, useRef } from "react"
-import { createPortal } from "react-dom"
-import { Plus, Pencil, Trash2, X, Repeat, Loader2 } from "lucide-react"
-import { toast } from "sonner"
-import {
-  useAddVariable,
-  useRemoveVariable,
-  useUpdateEstimator,
-  useUpdateVariable,
-} from "@/api/estimators.api"
-import {
-  idsToNames,
-  namesToIds,
-  namesToIdsPreservingIds,
-  type EstimatorIndex,
-} from "@/pages/flows/lib/expression-refs"
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Textarea } from "@/components/ui/textarea"
-import { Badge } from "@/components/ui/badge"
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select"
-import { Switch } from "@/components/ui/switch"
-import { Separator } from "@/components/ui/separator"
-import { cn } from "@/lib/utils"
-import type { Schemas } from "@/api/api.client"
-
-// ─── Panel state (ID-based — data derived from live query) ────────────────────
-
-export type PanelState =
-  | { mode: "add-step" }
-  | { mode: "step-details"; stepId: string }
-  | { mode: "add-field"; stepId: string }
-  | { mode: "edit-field"; fieldId: string; stepId: string }
-  | { mode: "estimator-details"; estimatorId: string }
-
-// ─── Field type helpers ───────────────────────────────────────────────────────
-
-type FieldType = Schemas.FieldConfigDto["type"]
-
-function defaultConfig(type: FieldType): Schemas.FieldConfigDto {
-  switch (type) {
-    case "text":
-      return { type: "text", max_length: 255 }
-    case "number":
-      return { type: "number", min: null, max: null }
-    case "date":
-      return { type: "date", min: "", max: "" }
-    case "boolean":
-      return { type: "boolean", default: false }
-    case "select":
-      return { type: "select", options: [] }
-  }
-}
-
-function labelToKey(label: string) {
-  return label
-    .toLowerCase()
-    .replace(/\s+/g, "_")
-    .replace(/[^a-z0-9_]/g, "")
-}
-
-// ─── Panel header ─────────────────────────────────────────────────────────────
-
-function PanelHeader({
-  title,
-  description,
-  onClose,
-  actions,
-}: {
-  title: string
-  description?: string
-  onClose: () => void
-  actions?: React.ReactNode
-}) {
-  return (
-    <div className="flex items-start gap-2 px-5 py-4 border-b shrink-0">
-      <div className="flex-1 min-w-0">
-        <h2 className="text-sm font-semibold leading-tight truncate">{title}</h2>
-        {description && (
-          <p className="text-xs text-muted-foreground mt-0.5 truncate">{description}</p>
-        )}
-      </div>
-      {actions}
-      <Button variant="ghost" size="icon" className="w-7 h-7 shrink-0" onClick={onClose}>
-        <X className="w-4 h-4" />
-      </Button>
-    </div>
-  )
-}
-
-// ─── Field config sub-form ────────────────────────────────────────────────────
-
-function FieldConfigForm({
-  config,
-  onChange,
-}: {
-  config: Schemas.FieldConfigDto
-  onChange: (config: Schemas.FieldConfigDto) => void
-}) {
-  if (config.type === "text") {
-    return (
-      <div className="flex flex-col gap-1.5">
-        <Label htmlFor="max_length">Max length</Label>
-        <Input
-          id="max_length"
-          type="number"
-          min={1}
-          value={config.max_length}
-          onChange={(e) => onChange({ type: "text", max_length: Number(e.target.value) })}
-        />
-      </div>
-    )
-  }
-
-  if (config.type === "number") {
-    return (
-      <div className="grid grid-cols-2 gap-3">
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="num_min">Min</Label>
-          <Input
-            id="num_min"
-            type="number"
-            placeholder="No limit"
-            value={config.min ?? ""}
-            onChange={(e) =>
-              onChange({ ...config, min: e.target.value === "" ? null : Number(e.target.value) })
-            }
-          />
-        </div>
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="num_max">Max</Label>
-          <Input
-            id="num_max"
-            type="number"
-            placeholder="No limit"
-            value={config.max ?? ""}
-            onChange={(e) =>
-              onChange({ ...config, max: e.target.value === "" ? null : Number(e.target.value) })
-            }
-          />
-        </div>
-      </div>
-    )
-  }
-
-  if (config.type === "date") {
-    return (
-      <div className="grid grid-cols-2 gap-3">
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="date_min">Min date</Label>
-          <Input
-            id="date_min"
-            type="date"
-            value={config.min}
-            onChange={(e) => onChange({ ...config, min: e.target.value })}
-          />
-        </div>
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="date_max">Max date</Label>
-          <Input
-            id="date_max"
-            type="date"
-            value={config.max}
-            onChange={(e) => onChange({ ...config, max: e.target.value })}
-          />
-        </div>
-      </div>
-    )
-  }
-
-  if (config.type === "select") {
-    return (
-      <div className="flex flex-col gap-1.5">
-        <Label>Options (one per line)</Label>
-        <Textarea
-          value={config.options.join("\n")}
-          onChange={(e) =>
-            onChange({ type: "select", options: e.target.value.split("\n").filter(Boolean) })
-          }
-          placeholder={"Option A\nOption B\nOption C"}
-          rows={4}
-        />
-      </div>
-    )
-  }
-
-  return null
-}
-
-// ─── Main panel ───────────────────────────────────────────────────────────────
-
-type Props = {
-  state: PanelState | null
-  step: Schemas.StepResponse | null
-  field: Schemas.FieldResponse | null
-  estimator: Schemas.EstimatorResponse | null
-  flowId: string
-  availableFieldKeys: string[]
-  otherEstimators: Array<{ id: string; name: string; variables: string[] }>
-  estimatorsIndex: EstimatorIndex
-  onClose: () => void
-  onAddStep: (data: { title: string; description: string }) => void
-  onUpdateStep: (stepId: string, data: Schemas.UpdateStepMetadataRequest) => void
-  onAddField: (stepId: string, data: { label: string; key: string; config: Schemas.FieldConfigDto }) => void
-  onEditField: (fieldId: string, stepId: string, data: { label: string; config: Schemas.FieldConfigDto }) => void
-  onDeleteField: (fieldId: string, stepId: string) => void
-  onOpenAddField: (stepId: string) => void
-  onOpenEditField: (fieldId: string, stepId: string) => void
-}
-
-export function FlowEditPanel({
-  state,
-  step,
-  field,
-  estimator,
-  flowId,
-  availableFieldKeys,
-  otherEstimators,
-  estimatorsIndex,
-  onClose,
-  onAddStep,
-  onUpdateStep,
-  onAddField,
-  onEditField,
-  onDeleteField,
-  onOpenAddField,
-  onOpenEditField,
-}: Props) {
-  return (
-    <div
-      className={cn(
-        "shrink-0 overflow-hidden transition-[width] duration-200 ease-in-out",
-        state ? "w-80" : "w-0",
-      )}
-    >
-      <div className="w-80 h-full border-l bg-background flex flex-col overflow-hidden">
-        {state?.mode === "add-step" && (
-          <AddStepForm onClose={onClose} onSubmit={onAddStep} />
-        )}
-        {state?.mode === "step-details" && step && (
-          <StepDetailsPanel
-            step={step}
-            onClose={onClose}
-            onUpdateStep={onUpdateStep}
-            onAddField={onOpenAddField}
-            onEditField={onOpenEditField}
-            onDeleteField={onDeleteField}
-          />
-        )}
-        {state?.mode === "add-field" && step && (
-          <AddFieldForm
-            stepId={step.id}
-            stepTitle={step.title}
-            onClose={onClose}
-            onSubmit={onAddField}
-          />
-        )}
-        {state?.mode === "edit-field" && field && state.stepId && (
-          <EditFieldForm
-            field={field}
-            stepId={state.stepId}
-            onClose={onClose}
-            onSubmit={onEditField}
-          />
-        )}
-        {state?.mode === "estimator-details" && estimator && (
-          <EstimatorDetailsPanel
-            key={estimator.id}
-            estimator={estimator}
-            flowId={flowId}
-            availableFieldKeys={availableFieldKeys}
-            otherEstimators={otherEstimators}
-            estimatorsIndex={estimatorsIndex}
-            onClose={onClose}
-          />
-        )}
-      </div>
-    </div>
-  )
-}
-
-// ─── Add Step Form ────────────────────────────────────────────────────────────
-
-function AddStepForm({
-  onClose,
-  onSubmit,
-}: {
-  onClose: () => void
-  onSubmit: (data: { title: string; description: string }) => void
-}) {
-  const [title, setTitle] = useState("")
-  const [description, setDescription] = useState("")
-
-  return (
-    <>
-      <PanelHeader title="Add step" description="Add a new step to this flow." onClose={onClose} />
-      <div className="flex flex-col gap-4 px-5 py-4 flex-1 overflow-y-auto">
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="step-title">Title *</Label>
-          <Input
-            id="step-title"
-            placeholder="e.g. Personal information"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-          />
-        </div>
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="step-desc">Description</Label>
-          <Textarea
-            id="step-desc"
-            placeholder="Describe this step…"
-            value={description}
-            onChange={(e) => setDescription(e.target.value)}
-            rows={3}
-          />
-        </div>
-      </div>
-      <div className="flex gap-2 px-5 py-4 border-t shrink-0">
-        <Button variant="outline" className="flex-1" onClick={onClose}>
-          Cancel
-        </Button>
-        <Button
-          className="flex-1"
-          disabled={!title.trim()}
-          onClick={() => {
-            onSubmit({ title: title.trim(), description: description.trim() })
-            onClose()
-          }}
-        >
-          Add step
-        </Button>
-      </div>
-    </>
-  )
-}
-
-// ─── Step Details Panel ───────────────────────────────────────────────────────
-
-const STEP_COLOR = "hsl(28, 85%, 55%)"
-
-function StepDetailsPanel({
-  step,
-  onClose,
-  onUpdateStep,
-  onAddField,
-  onEditField,
-  onDeleteField,
-}: {
-  step: Schemas.StepResponse
-  onClose: () => void
-  onUpdateStep: (stepId: string, data: Schemas.UpdateStepMetadataRequest) => void
-  onAddField: (stepId: string) => void
-  onEditField: (fieldId: string, stepId: string) => void
-  onDeleteField: (fieldId: string, stepId: string) => void
-}) {
-  const [editingTitle, setEditingTitle] = useState(false)
-  const [title, setTitle] = useState(step.title)
-
-  useEffect(() => {
-    setTitle(step.title)
-    setEditingTitle(false)
-  }, [step.id, step.title])
-
-  return (
-    <>
-      <div className="flex items-center gap-2 px-5 pt-4 pb-2 border-b shrink-0">
-        {editingTitle ? (
-          <Input
-            autoFocus
-            className="flex-1 !text-base font-semibold h-7 rounded-sm border border-border/60 bg-transparent px-2 py-0 shadow-none focus-visible:border-border focus-visible:ring-0"
-            style={{ color: STEP_COLOR }}
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-            onBlur={() => {
-              if (title.trim() && title.trim() !== step.title) {
-                onUpdateStep(step.id, { title: title.trim() })
-              }
-              setEditingTitle(false)
-            }}
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                if (title.trim() && title.trim() !== step.title) {
-                  onUpdateStep(step.id, { title: title.trim() })
-                }
-                setEditingTitle(false)
-              }
-              if (e.key === "Escape") {
-                setTitle(step.title)
-                setEditingTitle(false)
-              }
-            }}
-          />
-        ) : (
-          <button
-            className="flex-1 text-left text-base font-semibold truncate hover:opacity-80 transition-opacity cursor-text"
-            style={{ color: STEP_COLOR }}
-            onClick={() => setEditingTitle(true)}
-          >
-            {step.title}
-          </button>
-        )}
-        {!editingTitle && (
-          <Button variant="ghost" size="icon" className="shrink-0 h-7 w-7" onClick={onClose}>
-            <X className="w-4 h-4" />
-          </Button>
-        )}
-      </div>
-
-      {/* Repeatable configuration */}
-      <StepRepeatableConfig step={step} onUpdateStep={onUpdateStep} />
-
-      <div className="flex items-center justify-between px-5 py-3 border-b shrink-0">
-        <span className="text-xs font-medium text-muted-foreground uppercase tracking-wide">
-          Fields
-        </span>
-        <Button size="sm" variant="outline" className="gap-1.5 h-7 text-xs" onClick={() => onAddField(step.id)}>
-          <Plus className="w-3 h-3" />
-          Add field
-        </Button>
-      </div>
-
-      <div className="flex-1 overflow-y-auto">
-        {step.fields.length === 0 ? (
-          <div className="flex flex-col items-center justify-center gap-2 py-10 text-center px-5">
-            <p className="text-sm text-muted-foreground">No fields yet.</p>
-            <Button size="sm" variant="outline" className="gap-1.5" onClick={() => onAddField(step.id)}>
-              <Plus className="w-3.5 h-3.5" />
-              Add field
-            </Button>
-          </div>
-        ) : (
-          <ul className="flex flex-col divide-y">
-            {step.fields.map((field) => (
-              <li key={field.id} className="flex items-center gap-2 px-5 py-2.5">
-                <div className="flex-1 min-w-0">
-                  <p className="text-sm font-medium truncate">{field.label}</p>
-                  <p className="text-xs text-muted-foreground">{field.key}</p>
-                </div>
-                <Badge variant="secondary" className="shrink-0 text-xs">
-                  {field.config.type}
-                </Badge>
-                <div className="flex items-center gap-0.5 shrink-0">
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    className="w-6 h-6"
-                    onClick={() => onEditField(field.id, step.id)}
-                  >
-                    <Pencil className="w-3 h-3" />
-                  </Button>
-                  <AlertDialog>
-                    <AlertDialogTrigger asChild>
-                      <Button
-                        variant="ghost"
-                        size="icon"
-                        className="w-6 h-6 text-destructive hover:text-destructive"
-                      >
-                        <Trash2 className="w-3 h-3" />
-                      </Button>
-                    </AlertDialogTrigger>
-                    <AlertDialogContent>
-                      <AlertDialogHeader>
-                        <AlertDialogTitle>Delete field?</AlertDialogTitle>
-                        <AlertDialogDescription>
-                          "{field.label}" will be permanently deleted.
-                        </AlertDialogDescription>
-                      </AlertDialogHeader>
-                      <AlertDialogFooter>
-                        <AlertDialogCancel>Cancel</AlertDialogCancel>
-                        <AlertDialogAction
-                          className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-                          onClick={() => onDeleteField(field.id, step.id)}
-                        >
-                          Delete
-                        </AlertDialogAction>
-                      </AlertDialogFooter>
-                    </AlertDialogContent>
-                  </AlertDialog>
-                </div>
-              </li>
-            ))}
-          </ul>
-        )}
-      </div>
-
-    </>
-  )
-}
-
-// ─── Step Repeatable Config ──────────────────────────────────────────────────
-
-function StepRepeatableConfig({
-  step,
-  onUpdateStep,
-}: {
-  step: Schemas.StepResponse
-  onUpdateStep: (stepId: string, data: Schemas.UpdateStepMetadataRequest) => void
-}) {
-  const [isRepeatable, setIsRepeatable] = useState(step.is_repeatable)
-  const [repeatLabel, setRepeatLabel] = useState(step.repeat_label ?? "")
-  const [minRepeats, setMinRepeats] = useState(step.min_repeats)
-  const [maxRepeats, setMaxRepeats] = useState<number | "">(step.max_repeats ?? "")
-  const [validationError, setValidationError] = useState<string | null>(null)
-
-  // Sync local state when step changes (e.g. after server response)
-  useEffect(() => {
-    setIsRepeatable(step.is_repeatable)
-    setRepeatLabel(step.repeat_label ?? "")
-    setMinRepeats(step.min_repeats)
-    setMaxRepeats(step.max_repeats ?? "")
-    setValidationError(null)
-  }, [step.id, step.is_repeatable, step.repeat_label, step.min_repeats, step.max_repeats])
-
-  function validate(min: number, max: number | ""): string | null {
-    if (max !== "" && max < min) return "Max repeats must be greater than or equal to min."
-    if (min < 0) return "Min repeats cannot be negative."
-    return null
-  }
-
-  function handleToggle(checked: boolean) {
-    setIsRepeatable(checked)
-    if (!checked) {
-      setValidationError(null)
-      onUpdateStep(step.id, {
-        is_repeatable: false,
-        repeat_label: null,
-        min_repeats: 1,
-        max_repeats: null,
-      })
-    }
-  }
-
-  function handleSave() {
-    const error = validate(minRepeats, maxRepeats)
-    if (error) {
-      setValidationError(error)
-      return
-    }
-    setValidationError(null)
-    onUpdateStep(step.id, {
-      is_repeatable: true,
-      repeat_label: repeatLabel.trim() || null,
-      min_repeats: minRepeats,
-      max_repeats: maxRepeats === "" ? null : maxRepeats,
-    })
-  }
-
-  return (
-    <div className="px-5 py-3 border-b space-y-3">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <Repeat className="w-3.5 h-3.5 text-muted-foreground" />
-          <Label htmlFor="repeatable-toggle" className="text-xs font-medium">
-            Allow repetition
-          </Label>
-        </div>
-        <Switch
-          id="repeatable-toggle"
-          checked={isRepeatable}
-          onCheckedChange={handleToggle}
-        />
-      </div>
-
-      {isRepeatable && (
-        <div className="flex flex-col gap-3 pt-1">
-          <div className="flex flex-col gap-1.5">
-            <Label htmlFor="repeat-label" className="text-xs">Button label</Label>
-            <Input
-              id="repeat-label"
-              placeholder='e.g. "Add another room"'
-              value={repeatLabel}
-              onChange={(e) => setRepeatLabel(e.target.value)}
-            />
-          </div>
-
-          <div className="grid grid-cols-2 gap-3">
-            <div className="flex flex-col gap-1.5">
-              <Label htmlFor="min-repeats" className="text-xs">Min repeats</Label>
-              <Input
-                id="min-repeats"
-                type="number"
-                min={0}
-                value={minRepeats}
-                onChange={(e) => setMinRepeats(Number(e.target.value))}
-              />
-            </div>
-            <div className="flex flex-col gap-1.5">
-              <Label htmlFor="max-repeats" className="text-xs">Max repeats</Label>
-              <Input
-                id="max-repeats"
-                type="number"
-                min={0}
-                placeholder="No limit"
-                value={maxRepeats}
-                onChange={(e) => setMaxRepeats(e.target.value === "" ? "" : Number(e.target.value))}
-              />
-            </div>
-          </div>
-
-          {validationError && (
-            <p className="text-xs text-destructive">{validationError}</p>
-          )}
-
-          <Button size="sm" className="w-full" onClick={handleSave}>
-            Save repetition settings
-          </Button>
-        </div>
-      )}
-    </div>
-  )
-}
-
-// ─── Add Field Form ───────────────────────────────────────────────────────────
-
-function AddFieldForm({
-  stepId,
-  stepTitle,
-  onClose,
-  onSubmit,
-}: {
-  stepId: string
-  stepTitle: string
-  onClose: () => void
-  onSubmit: (stepId: string, data: { label: string; key: string; config: Schemas.FieldConfigDto }) => void
-}) {
-  const [label, setLabel] = useState("")
-  const [key, setKey] = useState("")
-  const [keyTouched, setKeyTouched] = useState(false)
-  const [config, setConfig] = useState<Schemas.FieldConfigDto>(defaultConfig("text"))
-
-  function handleLabelChange(value: string) {
-    setLabel(value)
-    if (!keyTouched) setKey(labelToKey(value))
-  }
-
-  function handleTypeChange(value: FieldType) {
-    setConfig(defaultConfig(value))
-  }
-
-  return (
-    <>
-      <PanelHeader title="Add field" description={`To "${stepTitle}"`} onClose={onClose} />
-      <div className="flex flex-col gap-4 px-5 py-4 flex-1 overflow-y-auto">
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="field-label">Label *</Label>
-          <Input
-            id="field-label"
-            placeholder="e.g. First name"
-            value={label}
-            onChange={(e) => handleLabelChange(e.target.value)}
-          />
-        </div>
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="field-key">Key *</Label>
-          <Input
-            id="field-key"
-            placeholder="e.g. first_name"
-            value={key}
-            onChange={(e) => {
-              setKeyTouched(true)
-              setKey(e.target.value)
-            }}
-          />
-        </div>
-        <div className="flex flex-col gap-1.5">
-          <Label>Type</Label>
-          <Select value={config.type} onValueChange={(v) => handleTypeChange(v as FieldType)}>
-            <SelectTrigger>
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="text">Text</SelectItem>
-              <SelectItem value="number">Number</SelectItem>
-              <SelectItem value="date">Date</SelectItem>
-              <SelectItem value="boolean">Boolean</SelectItem>
-              <SelectItem value="select">Select</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-        <FieldConfigForm config={config} onChange={setConfig} />
-      </div>
-      <div className="flex gap-2 px-5 py-4 border-t shrink-0">
-        <Button variant="outline" className="flex-1" onClick={onClose}>
-          Cancel
-        </Button>
-        <Button
-          className="flex-1"
-          disabled={!label.trim() || !key.trim()}
-          onClick={() => {
-            onSubmit(stepId, { label: label.trim(), key: key.trim(), config })
-            onClose()
-          }}
-        >
-          Add field
-        </Button>
-      </div>
-    </>
-  )
-}
-
-// ─── Edit Field Form ──────────────────────────────────────────────────────────
-
-function EditFieldForm({
-  field,
-  stepId,
-  onClose,
-  onSubmit,
-}: {
-  field: Schemas.FieldResponse
-  stepId: string
-  onClose: () => void
-  onSubmit: (fieldId: string, stepId: string, data: { label: string; config: Schemas.FieldConfigDto }) => void
-}) {
-  const [label, setLabel] = useState(field.label)
-  const [config, setConfig] = useState<Schemas.FieldConfigDto>(field.config)
-
-  useEffect(() => {
-    setLabel(field.label)
-    setConfig(field.config)
-  }, [field.id])
-
-  return (
-    <>
-      <PanelHeader
-        title="Edit field"
-        description={`${field.label} — ${field.config.type}`}
-        onClose={onClose}
-      />
-      <div className="flex flex-col gap-4 px-5 py-4 flex-1 overflow-y-auto">
-        <div className="flex flex-col gap-1.5">
-          <Label htmlFor="edit-field-label">Label *</Label>
-          <Input
-            id="edit-field-label"
-            value={label}
-            onChange={(e) => setLabel(e.target.value)}
-          />
-        </div>
-        <FieldConfigForm config={config} onChange={setConfig} />
-      </div>
-      <div className="flex gap-2 px-5 py-4 border-t shrink-0">
-        <Button variant="outline" className="flex-1" onClick={onClose}>
-          Cancel
-        </Button>
-        <Button
-          className="flex-1"
-          disabled={!label.trim()}
-          onClick={() => {
-            onSubmit(field.id, stepId, { label: label.trim(), config })
-            onClose()
-          }}
-        >
-          Save
-        </Button>
-      </div>
-    </>
-  )
-}
-
-// ─── Estimator Details Panel ─────────────────────────────────────────────────
-
-const ROSE = "hsl(330, 80%, 60%)"
-const AGG_FUNCTIONS = ["SUM", "AVG", "COUNT_ITER"] as const
-
-function EstimatorDetailsPanel({
-  estimator,
-  flowId,
-  availableFieldKeys,
-  otherEstimators,
-  estimatorsIndex,
-  onClose,
-}: {
-  estimator: Schemas.EstimatorResponse
-  flowId: string
-  availableFieldKeys: string[]
-  otherEstimators: Array<{ id: string; name: string; variables: string[] }>
-  estimatorsIndex: EstimatorIndex
-  onClose: () => void
-}) {
-  const [editingName, setEditingName] = useState(false)
-  // Display form uses spaces in place of underscores; commit replaces them back.
-  const [name, setName] = useState(estimator.name.replace(/_/g, " "))
-
-  useEffect(() => {
-    setName(estimator.name.replace(/_/g, " "))
-    setEditingName(false)
-  }, [estimator.id, estimator.name])
-
-  const updateEstimator = useUpdateEstimator(flowId, estimator.id)
-  const addVariable = useAddVariable(flowId, estimator.id)
-  const updateVariable = useUpdateVariable(flowId, estimator.id)
-  const removeVariable = useRemoveVariable(flowId, estimator.id)
-
-  const commitName = (next: string) => {
-    const trimmed = next.trim().replace(/\s+/g, "_")
-    if (!trimmed) return
-    if (trimmed === estimator.name) return
-    if (!/^[A-Za-z0-9_]+$/.test(trimmed)) {
-      toast.error("Estimator name can only contain letters, digits, spaces and underscores")
-      setName(estimator.name.replace(/_/g, " "))
-      return
-    }
-    // Prevent duplicate names: would break cross-ref name→id lookup
-    if (otherEstimators.some((e) => e.name === trimmed)) {
-      toast.error(`An estimator named "${trimmed.replace(/_/g, " ")}" already exists`)
-      setName(estimator.name.replace(/_/g, " "))
-      return
-    }
-    updateEstimator.mutate(
-      { path: { estimator_id: estimator.id }, body: { name: trimmed } },
-      { onError: (err) => toast.error(`Rename failed: ${err.message}`) },
-    )
-  }
-
-  const handleAddVariable = () => {
-    addVariable.mutate(
-      {
-        path: { estimator_id: estimator.id },
-        body: { name: "new_var", expression: "0", description: null },
-      },
-      { onError: (err) => toast.error(`Add variable failed: ${err.message}`) },
-    )
-  }
-
-  const handleUpdateVariable = (variableId: string, patch: Partial<Schemas.VariableResponse>) => {
-    const body: Schemas.UpdateVariableRequest = {
-      name: patch.name,
-      expression: patch.expression,
-      description: patch.description ?? null,
-    }
-    updateVariable.mutate(
-      { path: { variable_id: variableId }, body },
-      { onError: (err) => toast.error(`Update failed: ${err.message}`) },
-    )
-  }
-
-  const handleDeleteVariable = (variableId: string) => {
-    removeVariable.mutate(
-      { path: { variable_id: variableId } },
-      { onError: (err) => toast.error(`Delete failed: ${err.message}`) },
-    )
-  }
-
-  return (
-    <>
-      <div className="flex items-center gap-2 px-5 pt-4 pb-2 border-b shrink-0">
-        {editingName ? (
-          <Input
-            autoFocus
-            className="flex-1 !text-base font-semibold h-7 rounded-sm border border-border/60 bg-transparent px-2 py-0 shadow-none focus-visible:border-border focus-visible:ring-0"
-            style={{ color: ROSE }}
-            value={name}
-            onChange={(e) => {
-              // Allow letters, digits, spaces and underscores; strip everything else.
-              const allowed = e.target.value.replace(/[^A-Za-z0-9_ ]/g, "")
-              setName(allowed)
-            }}
-            onBlur={() => {
-              commitName(name)
-              setEditingName(false)
-            }}
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                commitName(name)
-                setEditingName(false)
-              }
-              if (e.key === "Escape") {
-                setName(estimator.name.replace(/_/g, " "))
-                setEditingName(false)
-              }
-            }}
-          />
-        ) : (
-          <button
-            className="flex-1 text-left text-base font-semibold truncate hover:opacity-80 transition-opacity cursor-text"
-            style={{ color: ROSE }}
-            onClick={() => setEditingName(true)}
-          >
-            {estimator.name.replace(/_/g, " ")}
-          </button>
-        )}
-        {!editingName && (
-          <Button variant="ghost" size="icon" className="shrink-0 h-7 w-7" onClick={onClose}>
-            <X className="w-4 h-4" />
-          </Button>
-        )}
-      </div>
-
-      <div className="flex flex-col gap-3 px-5 py-4 flex-1 overflow-y-auto pb-72">
-        <p className="text-xs text-muted-foreground">
-          Variables are evaluated in dependency order. Reference fields with <code className="font-mono bg-muted px-1 rounded">@field_key</code>.
-        </p>
-
-        {estimator.variables.map((v) => (
-          <VariableCard
-            key={v.id}
-            variable={v}
-            ownEstimatorName={estimator.name}
-            ownEstimatorVariables={estimator.variables
-              .filter((ov) => ov.id !== v.id)
-              .map((ov) => ov.name)}
-            availableFieldKeys={availableFieldKeys}
-            otherEstimators={otherEstimators}
-            estimatorsIndex={estimatorsIndex}
-            onUpdate={handleUpdateVariable}
-            onDelete={handleDeleteVariable}
-          />
-        ))}
-
-        <Button
-          variant="outline"
-          size="sm"
-          className="w-full border-dashed"
-          onClick={handleAddVariable}
-          disabled={addVariable.isPending}
-        >
-          {addVariable.isPending ? (
-            <Loader2 className="w-3.5 h-3.5 mr-1.5 animate-spin" />
-          ) : (
-            <Plus className="w-3.5 h-3.5 mr-1.5" />
-          )}
-          Add variable
-        </Button>
-      </div>
-    </>
-  )
-}
-
-// ─── Variable Card (inline editing with local drafts) ────────────────────────
-
-function VariableCard({
-  variable,
-  ownEstimatorName,
-  ownEstimatorVariables,
-  availableFieldKeys,
-  otherEstimators,
-  estimatorsIndex,
-  onUpdate,
-  onDelete,
-}: {
-  variable: Schemas.VariableResponse
-  ownEstimatorName: string
-  ownEstimatorVariables: string[]
-  availableFieldKeys: string[]
-  otherEstimators: Array<{ id: string; name: string; variables: string[] }>
-  estimatorsIndex: EstimatorIndex
-  onUpdate: (variableId: string, patch: Partial<Schemas.VariableResponse>) => void
-  onDelete: (variableId: string) => void
-}) {
-  const [expanded, setExpanded] = useState(!variable.expression)
-  const [showSuggestions, setShowSuggestions] = useState(false)
-  const [suggestionFilter, setSuggestionFilter] = useState("")
-  const exprRef = useRef<HTMLTextAreaElement>(null)
-
-  // Expression is stored with `@#<uuid>.var`; display with `@Name.var`
-  const displayExpression = idsToNames(variable.expression, estimatorsIndex)
-
-  // Local drafts — sync from prop when it changes (e.g. after server update or switching variable)
-  const [nameDraft, setNameDraft] = useState(variable.name)
-  const [exprDraft, setExprDraft] = useState(displayExpression)
-  const [descDraft, setDescDraft] = useState(variable.description)
-  // Remember which estimator id was chosen for each `name.var` via autocomplete,
-  // so free-form editing doesn't silently reroute to the first same-named estimator.
-  const pickedIdsRef = useRef<Map<string, string>>(new Map())
-
-  useEffect(() => {
-    setNameDraft(variable.name)
-    setExprDraft(idsToNames(variable.expression, estimatorsIndex))
-    setDescDraft(variable.description)
-    pickedIdsRef.current = new Map()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [variable.id, variable.name, variable.expression, variable.description])
-
-  // Dropdown positioning: track the textarea's viewport position so the
-  // dropdown can be rendered via portal (escaping all overflow clips).
-  const [dropdownPos, setDropdownPos] = useState<{
-    top: number
-    left: number
-    width: number
-    placement: "below" | "above"
-  } | null>(null)
-
-  useEffect(() => {
-    if (!showSuggestions) {
-      setDropdownPos(null)
-      return
-    }
-    const el = exprRef.current
-    if (!el) return
-
-    function updatePos() {
-      if (!el) return
-      const rect = el.getBoundingClientRect()
-      const spaceBelow = window.innerHeight - rect.bottom
-      const spaceAbove = rect.top
-      const maxHeight = 240
-      const placement: "below" | "above" =
-        spaceBelow >= Math.min(maxHeight, 150) || spaceBelow >= spaceAbove
-          ? "below"
-          : "above"
-      setDropdownPos({
-        top: placement === "below" ? rect.bottom + 4 : rect.top - 4,
-        left: rect.left,
-        width: rect.width,
-        placement,
-      })
-    }
-
-    updatePos()
-    window.addEventListener("scroll", updatePos, true)
-    window.addEventListener("resize", updatePos)
-    return () => {
-      window.removeEventListener("scroll", updatePos, true)
-      window.removeEventListener("resize", updatePos)
-    }
-  }, [showSuggestions])
-
-  const filterLower = suggestionFilter.toLowerCase()
-  type Suggestion = {
-    label: string
-    insert: string
-    group: string
-    groupLabel?: string
-    isEstimator?: boolean
-    estimatorId?: string
-  }
-  const suggestions: Suggestion[] = [
-    // Fields
-    ...availableFieldKeys
-      .filter((k) => k.toLowerCase().includes(filterLower))
-      .map((k) => ({ label: `@${k}`, insert: `@${k}`, group: "Fields" })),
-    // Functions
-    ...AGG_FUNCTIONS
-      .filter((fn) => fn.toLowerCase().includes(filterLower))
-      .map((fn) => ({ label: `${fn}(@...)`, insert: `${fn}(@)`, group: "Functions" })),
-    // Variables of the current estimator (bare references)
-    ...ownEstimatorVariables
-      .filter((v) => v.toLowerCase().includes(filterLower))
-      .map((v) => ({
-        label: `@${v}`,
-        insert: `@${v}`,
-        group: ownEstimatorName.replace(/_/g, " "),
-        isEstimator: true,
-      })),
-    // Variables of other estimators (cross-references).
-    // Group key uses estimator id so duplicates stay separate.
-    ...otherEstimators.flatMap((est) => {
-      const displayName = est.name.replace(/_/g, " ")
-      return est.variables
-        .filter((v) =>
-          v.toLowerCase().includes(filterLower) ||
-          `${est.name}.${v}`.toLowerCase().includes(filterLower) ||
-          `${displayName}.${v}`.toLowerCase().includes(filterLower),
-        )
-        .map((v) => ({
-          label: `@${displayName}.${v}`,
-          insert: `@${est.name}.${v}`,
-          // Unique key per estimator instance (id embedded, hidden at render)
-          group: `${displayName}\u0000${est.id}`,
-          groupLabel: displayName,
-          isEstimator: true,
-          estimatorId: est.id,
-        }))
-    }),
-  ]
-
-  function insertSuggestion(s: Suggestion) {
-    const el = exprRef.current
-    if (!el) return
-    const pos = el.selectionStart ?? el.value.length
-    const before = el.value.substring(0, pos)
-    const atPos = before.lastIndexOf("@")
-    const start = atPos >= 0 ? atPos : pos
-    const after = el.value.substring(pos)
-    const newVal = el.value.substring(0, start) + s.insert + after
-    setExprDraft(newVal)
-    setShowSuggestions(false)
-
-    // Remember the estimator id picked for this name.var pair so subsequent
-    // commits (blur / typing changes) resolve to the correct estimator
-    // even when two share a name.
-    if (s.estimatorId) {
-      const est = estimatorsIndex.find((e) => e.id === s.estimatorId)
-      const varName = s.insert.replace(/^@[^.]+\./, "")
-      if (est) {
-        pickedIdsRef.current.set(`${est.name}.${varName}`, s.estimatorId)
-      }
-    }
-
-    // Build stored form using the preserving translator + recent picks.
-    const stored = namesToIdsPreservingIds(
-      newVal,
-      estimatorsIndex,
-      variable.expression,
-      pickedIdsRef.current,
-    )
-    onUpdate(variable.id, { expression: stored })
-
-    requestAnimationFrame(() => {
-      el.focus()
-      const cursorPos = start + s.insert.length
-      el.setSelectionRange(cursorPos, cursorPos)
-    })
-  }
-
-  function handleExpressionChange(value: string) {
-    setExprDraft(value)
-    const el = exprRef.current
-    if (el) {
-      const pos = el.selectionStart ?? value.length
-      const before = value.substring(0, pos)
-      const atMatch = before.match(/@([A-Za-z0-9_.]*)$/)
-      if (atMatch) {
-        setSuggestionFilter(atMatch[1])
-        setShowSuggestions(true)
-      } else {
-        setShowSuggestions(false)
-      }
-    }
-  }
-
-  // Commit helpers — only fire mutation if value actually changed
-  const commitName = () => {
-    const trimmed = nameDraft.trim()
-    if (!trimmed || trimmed === variable.name) {
-      setNameDraft(variable.name)
-      return
-    }
-    onUpdate(variable.id, { name: trimmed })
-  }
-  const commitExpr = () => {
-    // Backend requires non-empty expression
-    if (!exprDraft.trim()) {
-      setExprDraft(idsToNames(variable.expression, estimatorsIndex))
-      return
-    }
-    // Translate name-based refs to ID-based, preferring the ids already
-    // present in the current storage expression so we don't reroute an
-    // edge to a different same-named estimator.
-    const stored = namesToIdsPreservingIds(
-      exprDraft,
-      estimatorsIndex,
-      variable.expression,
-      pickedIdsRef.current,
-    )
-    if (stored === variable.expression) return
-    onUpdate(variable.id, { expression: stored })
-  }
-  const commitDesc = () => {
-    if (descDraft === variable.description) return
-    onUpdate(variable.id, { description: descDraft })
-  }
-
-  return (
-    <div className="rounded-md border border-border/60 overflow-hidden">
-      {/* Header row */}
-      <div className="flex items-center gap-1.5 px-3 py-2 bg-muted/30">
-        <button
-          className="flex-1 text-left"
-          onClick={() => setExpanded((p) => !p)}
-        >
-          <span className="text-sm font-mono font-semibold" style={{ color: ROSE }}>
-            {variable.name}
-          </span>
-          {!expanded && variable.expression && (
-            <span className="text-xs text-muted-foreground ml-2 font-mono">
-              = {variable.expression.length > 20 ? variable.expression.slice(0, 20) + "..." : variable.expression}
-            </span>
-          )}
-        </button>
-        <AlertDialog>
-          <AlertDialogTrigger asChild>
-            <Button variant="ghost" size="icon" className="h-6 w-6 text-destructive hover:text-destructive shrink-0">
-              <Trash2 className="w-3 h-3" />
-            </Button>
-          </AlertDialogTrigger>
-          <AlertDialogContent>
-            <AlertDialogHeader>
-              <AlertDialogTitle>Delete variable?</AlertDialogTitle>
-              <AlertDialogDescription>
-                "{variable.name}" will be permanently deleted.
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-              <AlertDialogCancel>Cancel</AlertDialogCancel>
-              <AlertDialogAction
-                className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-                onClick={() => onDelete(variable.id)}
-              >
-                Delete
-              </AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
-      </div>
-
-      {/* Expanded editor */}
-      {expanded && (
-        <div className="px-3 py-2.5 space-y-2.5 border-t border-border/40">
-          {/* Name */}
-          <div className="flex flex-col gap-1">
-            <Label className="text-xs">Name</Label>
-            <Input
-              className="h-7 text-sm font-mono"
-              value={nameDraft}
-              onChange={(e) => setNameDraft(e.target.value)}
-              onBlur={commitName}
-              onKeyDown={(e) => {
-                if (e.key === "Enter") (e.target as HTMLInputElement).blur()
-                if (e.key === "Escape") {
-                  setNameDraft(variable.name)
-                  ;(e.target as HTMLInputElement).blur()
-                }
-              }}
-            />
-          </div>
-
-          {/* Expression */}
-          <div className="flex flex-col gap-1 relative">
-            <Label className="text-xs">Expression</Label>
-            <Textarea
-              ref={exprRef}
-              className="text-sm font-mono min-h-[60px] resize-none"
-              placeholder="e.g. @surface * @prix_unitaire * 1.2"
-              value={exprDraft}
-              onChange={(e) => handleExpressionChange(e.target.value)}
-              onFocus={() => {
-                if (!exprDraft) setShowSuggestions(true)
-              }}
-              onBlur={() => {
-                // Delay to allow click on suggestion
-                setTimeout(() => {
-                  setShowSuggestions(false)
-                  commitExpr()
-                }, 200)
-              }}
-            />
-            {showSuggestions && suggestions.length > 0 && dropdownPos &&
-              createPortal(
-                <div
-                  style={{
-                    position: "fixed",
-                    top: dropdownPos.placement === "below" ? dropdownPos.top : undefined,
-                    bottom:
-                      dropdownPos.placement === "above"
-                        ? window.innerHeight - dropdownPos.top
-                        : undefined,
-                    left: dropdownPos.left,
-                    width: dropdownPos.width,
-                    zIndex: 9999,
-                  }}
-                  className="max-h-60 overflow-y-auto rounded-md border bg-popover shadow-md"
-                >
-                  {suggestions.map((s, i) => {
-                    const prevGroup = i > 0 ? suggestions[i - 1].group : null
-                    const showGroupLabel = s.group !== prevGroup
-                    return (
-                      <div key={`${s.group}-${s.insert}`}>
-                        {showGroupLabel && (
-                          <div className="px-3 pt-2 pb-1 text-[10px] font-semibold text-muted-foreground uppercase tracking-wider">
-                            {s.groupLabel ?? s.group}
-                          </div>
-                        )}
-                        <button
-                          className={cn(
-                            "w-full text-left px-3 py-1.5 text-xs font-mono hover:bg-accent transition-colors",
-                            s.isEstimator && "text-[hsl(330,70%,60%)]",
-                          )}
-                          onMouseDown={(e) => {
-                            e.preventDefault()
-                            insertSuggestion(s)
-                          }}
-                        >
-                          {s.label}
-                        </button>
-                      </div>
-                    )
-                  })}
-                </div>,
-                document.body,
-              )}
-          </div>
-
-          {/* Description */}
-          <div className="flex flex-col gap-1">
-            <Label className="text-xs">Description</Label>
-            <Input
-              className="h-7 text-sm"
-              placeholder="Optional"
-              value={descDraft}
-              onChange={(e) => setDescDraft(e.target.value)}
-              onBlur={commitDesc}
-            />
-          </div>
-        </div>
-      )}
-    </div>
-  )
-}
